AYS-246 | Admin Registration Application Mappers Have Been Moved to `user.mapper` Package
