+++ src/main/java/org/ays/user/model/mapper/AdminRegistrationApplicationToAdminRegistrationApplicationsResponseMapper.java
@@ -1,0 +1,29 @@
+package org.ays.user.model.mapper;
+
+import org.ays.common.model.mapper.BaseMapper;
+import org.ays.user.model.AdminRegistrationApplication;
+import org.ays.user.model.dto.response.AdminRegistrationApplicationsResponse;
+import org.mapstruct.Mapper;
+import org.mapstruct.factory.Mappers;
+
+
+/**
+ * AdminUserRegisterApplicationToAdminUserRegisterApplicationsResponse is an interface that defines the mapping between an {@link AdminRegistrationApplication} and an {@link AdminRegistrationApplicationsResponse}.
+ * This interface uses the MapStruct annotation @Mapper to generate an implementation of this interface at compile-time.
+ * <p>The class provides a static method {@code initialize()} that returns an instance of the generated mapper implementation.
+ * <p>The interface extends the MapStruct interface {@link BaseMapper}, which defines basic mapping methods.
+ * The interface adds no additional mapping methods, but simply defines the types to be used in the mapping process.
+ */
+@Mapper
+public interface AdminRegistrationApplicationToAdminRegistrationApplicationsResponseMapper extends BaseMapper<AdminRegistrationApplication, AdminRegistrationApplicationsResponse> {
+
+    /**
+     * Initializes the mapper.
+     *
+     * @return the initialized mapper object.
+     */
+    static AdminRegistrationApplicationToAdminRegistrationApplicationsResponseMapper initialize() {
+        return Mappers.getMapper(AdminRegistrationApplicationToAdminRegistrationApplicationsResponseMapper.class);
+    }
+
+}
