+++ src/main/java/org/ays/user/service/impl/AdminRegistrationApplicationServiceImpl.java
@@ -1,8 +1,6 @@
 package org.ays.user.service.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.ays.admin_user.model.mapper.AdminRegistrationApplicationCreateRequestToAdminRegistrationApplicationEntityMapper;
-import org.ays.admin_user.model.mapper.AdminRegistrationApplicationEntityToAdminRegistrationApplicationMapper;
 import org.ays.common.model.AysPage;
 import org.ays.institution.repository.InstitutionRepository;
 import org.ays.institution.util.exception.AysInstitutionNotExistException;
@@ -13,6 +11,8 @@ import org.ays.user.model.dto.request.AdminRegistrationApplicationRejectRequest;
 import org.ays.user.model.entity.AdminRegistrationApplicationEntity;
 import org.ays.user.model.entity.UserEntityV2;
 import org.ays.user.model.enums.AdminRegistrationApplicationStatus;
+import org.ays.user.model.mapper.AdminRegistrationApplicationCreateRequestToAdminRegistrationApplicationEntityMapper;
+import org.ays.user.model.mapper.AdminRegistrationApplicationEntityToAdminRegistrationApplicationMapper;
 import org.ays.user.repository.AdminRegistrationApplicationRepository;
 import org.ays.user.repository.UserRepositoryV2;
 import org.ays.user.service.AdminRegistrationApplicationService;
