+++ src/main/java/org/ays/user/service/impl/AdminRegistrationCompleteServiceImpl.java
@@ -2,7 +2,6 @@ package org.ays.user.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.ays.admin_user.model.mapper.AdminRegistrationApplicationCompleteRequestToUserEntityMapper;
 import org.ays.common.model.dto.request.AysPhoneNumberRequest;
 import org.ays.user.model.dto.request.AdminRegistrationApplicationCompleteRequest;
 import org.ays.user.model.entity.AdminRegistrationApplicationEntity;
@@ -10,6 +9,7 @@ import org.ays.user.model.entity.PermissionEntity;
 import org.ays.user.model.entity.RoleEntity;
 import org.ays.user.model.entity.UserEntityV2;
 import org.ays.user.model.enums.AdminRegistrationApplicationStatus;
+import org.ays.user.model.mapper.AdminRegistrationApplicationCompleteRequestToUserEntityMapper;
 import org.ays.user.repository.AdminRegistrationApplicationRepository;
 import org.ays.user.repository.PermissionRepository;
 import org.ays.user.repository.RoleRepository;
