+++ src/test/java/org/ays/admin_user/controller/AdminRegistrationApplicationControllerTest.java
@@ -7,11 +7,6 @@ import org.ays.admin_user.model.dto.request.AdminRegistrationApplicationCreateRe
 import org.ays.admin_user.model.dto.request.AdminRegistrationApplicationListRequestBuilder;
 import org.ays.admin_user.model.dto.request.AdminRegistrationApplicationRejectRequestBuilder;
 import org.ays.admin_user.model.entity.AdminRegisterApplicationEntityBuilder;
-import org.ays.admin_user.model.mapper.AdminRegistrationApplicationEntityToAdminRegistrationApplicationMapper;
-import org.ays.admin_user.model.mapper.AdminRegistrationApplicationToAdminRegistrationApplicationCreateResponseMapper;
-import org.ays.admin_user.model.mapper.AdminRegistrationApplicationToAdminRegistrationApplicationResponseMapper;
-import org.ays.admin_user.model.mapper.AdminRegistrationApplicationToAdminRegistrationApplicationSummaryResponseMapper;
-import org.ays.admin_user.model.mapper.AdminRegistrationApplicationToAdminRegistrationApplicationsResponseMapper;
 import org.ays.common.model.AysPage;
 import org.ays.common.model.dto.request.AysPhoneNumberRequest;
 import org.ays.common.model.dto.request.AysPhoneNumberRequestBuilder;
@@ -34,6 +29,11 @@ import org.ays.user.model.dto.response.AdminRegistrationApplicationSummaryRespon
 import org.ays.user.model.dto.response.AdminRegistrationApplicationsResponse;
 import org.ays.user.model.entity.AdminRegistrationApplicationEntity;
 import org.ays.user.model.enums.AdminRegistrationApplicationStatus;
+import org.ays.user.model.mapper.AdminRegistrationApplicationEntityToAdminRegistrationApplicationMapper;
+import org.ays.user.model.mapper.AdminRegistrationApplicationToAdminRegistrationApplicationCreateResponseMapper;
+import org.ays.user.model.mapper.AdminRegistrationApplicationToAdminRegistrationApplicationResponseMapper;
+import org.ays.user.model.mapper.AdminRegistrationApplicationToAdminRegistrationApplicationSummaryResponseMapper;
+import org.ays.user.model.mapper.AdminRegistrationApplicationToAdminRegistrationApplicationsResponseMapper;
 import org.ays.user.service.AdminRegistrationApplicationService;
 import org.ays.user.service.AdminRegistrationCompleteService;
 import org.ays.util.AysMockMvcRequestBuilders;
